Use lax parsing & report warning
This allows parsing to continue on error and catch more errors. It also behaves closer to how Liquid templates are rendered in prod.
